Move Controller to its own package and define controller from fxml directly
